Update README.md
Task-Track-Duration project ,this project contains two web pages one is "Add Task" and other one is "Task Record". In add task page we will add tasks and in in task record page we will able to see all the records which has been added in the database. 
WORKING:-
First of all we will add task as much as we want for a  different schedule jobs and the task will automatically gets deleted from the database after the completion of a given duration from the given current time .